backend/requirements.txt
selenium
beautifulsoup4
pymongo
fastapi
uvicorn
python-dotenv
webdriver-manager
apscheduler
requests
